feat(hackaton-refactor): ensure tests pass
